+++ tests/t_gp_growth.py
@@ -2,7 +2,7 @@ import time
 from collections import defaultdict
 
 import pysubgroup as ps
-from tests.DataSets import get_credit_data
+from pysubgroup.datasets import get_credit_data
 
 data = get_credit_data()
 # warnings.filterwarnings("error")