+++ tests/test_ChiSquaredQF.py
@@ -1,7 +1,7 @@
 from timeit import default_timer as timer
 
 import pysubgroup as ps
-from tests.DataSets import get_credit_data
+from pysubgroup.datasets import get_credit_data
 
 if __name__ == "__main__":
     data = get_credit_data()