+++ tests/test_SubgroupDiscoveryResult.py
@@ -1,8 +1,9 @@
 import unittest
 
+from algorithms_testing import TestAlgorithmsBase
+
 import pysubgroup as ps
-from tests.algorithms_testing import TestAlgorithmsBase
-from tests.DataSets import get_credit_data
+from pysubgroup.datasets import get_credit_data
 
 show_plots = False