+++ tests/test_algorithms.py
@@ -1,7 +1,7 @@
 import unittest
 
 import pysubgroup as ps
-from tests.DataSets import get_credit_data
+from pysubgroup.datasets import get_credit_data
 
 
 class TestAlgorithms(unittest.TestCase):