+++ tests/test_fi_target.py
@@ -1,10 +1,10 @@
 import unittest
 
 import numpy as np
+from algorithms_testing import TestAlgorithmsBase
 
 import pysubgroup as ps
-from tests.algorithms_testing import TestAlgorithmsBase
-from tests.DataSets import get_titanic_data
+from pysubgroup.datasets import get_titanic_data
 
 # data=get_titanic_data()
 # search_space = ps.create_selectors(data, ignore="survived")