+++ tests/test_qf_input.py
@@ -3,7 +3,7 @@ import unittest
 import numpy as np
 
 import pysubgroup as ps
-from tests.DataSets import get_credit_data
+from pysubgroup.datasets import get_credit_data
 
 
 class TestQFInputs(unittest.TestCase):