+++ CHANGELOG.md
@@ -85,7 +85,7 @@ This update prepares pysubgroup for a better future. To do so we had to break ba
 - SimpleSearch algorithm
 - DFS (Depth first search) using a representation for StandardQF
 - tests
-  - access to datasets for testing is provided through tests.DataSets class
+  - access to datasets for testing is provided through DataSets class
   - tests for selector classes (NominalSelector, NumericSelector)
     - \_\_eq\_\_
     - \_\_lt\_\_